ingest/requirements.txt
mysql-connector-python
smbus2
gsmHat
scapy